v如果被使用了不能删除和修改

+++ smarttt-evaluation/src/main/java/com/example/smartttevaluation/service/impl/ValueTagServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.smartttevaluation.service.impl;
 
+import cn.hutool.core.collection.CollectionUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.json.JSON;
 import cn.hutool.json.JSONUtil;
@@ -13,6 +14,7 @@ import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
 import java.time.LocalDateTime;
+import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -36,6 +38,14 @@ public class ValueTagServiceImpl implements ValueTagService {
 
     @Override
     public Boolean deleteValueTagByTypeId(List<Long> typeIdList) {
+        // 根据typeId查询所有的valueTag
+        List<ValueTag> valueTagList = valueTagMapper.getValueTagByTypeIdList(typeIdList);
+        if(CollectionUtil.isEmpty(valueTagList)) return true;
+        for (ValueTag valueTag : valueTagList){
+            if(valueTag.getCalCount()>0){
+                throw new RuntimeException("该valueTag:"+valueTag.getName()+"被使用，不能删除");
+            }
+        }
         return valueTagMapper.deleteValueTagByTypeId(typeIdList);
     }
 
@@ -57,12 +67,28 @@ public class ValueTagServiceImpl implements ValueTagService {
 
     @Override
     public Boolean update(ValueTag valueTag) {
+        List<ValueTag> valueTagList = valueTagMapper.getValueTagByIdList(Arrays.asList(valueTag.getId()));
+        if(CollectionUtil.isNotEmpty(valueTagList)){
+            ValueTag tag = valueTagList.get(0);
+            if(tag.getCalCount()>0){
+                throw new RuntimeException("该valueTag:"+tag.getName()+"被使用，不能修改");
+            }
+        }
         valueTag.setCreateTime(LocalDateTime.now());
         return valueTagMapper.update(valueTag);
     }
 
     @Override
     public Boolean delete(List<Long> idList) {
+        // 根据id获取整个记录信息
+        List<ValueTag> valueTagList = valueTagMapper.getValueTagByIdList(idList);
+        if(CollectionUtil.isEmpty(valueTagList)) return true;
+
+        for (ValueTag valueTag : valueTagList){
+            if(valueTag.getCalCount()>0){
+                throw new RuntimeException("该valueTag:"+valueTag.getName()+"被使用，不能删除");
+            }
+        }
         return valueTagMapper.delete(idList);
     }
 
@@ -80,6 +106,21 @@ public class ValueTagServiceImpl implements ValueTagService {
     public Boolean bindKwaAndValue(KwaBindReq kwaBindReq) {
         JSON parse = JSONUtil.parse(kwaBindReq.getKwaValueTagBindReqList());
         return valueTagMapper.bindKwaAndValue(kwaBindReq.getKwaId(), parse.toString());
+    }
+
+    @Override
+    public String getValueTagListJson(String kwaId) {
+        return valueTagMapper.getValueTagListJson(kwaId);
+    }
+
+    @Override
+    public void decreaseValueTagCalCount(List<Long> ids) {
+        valueTagMapper.decreaseValueTagCalCount(ids);
+    }
+
+    @Override
+    public void increaseValueTagCalCount(List<Long> newIds) {
+        valueTagMapper.increaseValueTagCalCount(newIds);
     }
 
 }
